2022/Hack The Box - Cyber Apocalypse/sources/web_amidst_us/challenge/requirements.txt
wheel
Pillow==8.4.0